Separate all remaining property-conversion tests
